connectors/binance: move URM and constants
- Move urm.py to connectors/binance/urm.py
- Move constants.py to connectors/binance/config.py (rename per architecture)
- Update all imports to use connectors namespace
- Add shim in providers/binance/urm.py for backward compatibility
- Update provider registration to import URM from connectors
